--- bet-it/bank/src/main/java/ch/unisg/BankApplication.java
@@ -1,20 +1,0 @@
-package ch.unisg;
-
-import io.camunda.zeebe.spring.client.EnableZeebeClient;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@EnableZeebeClient
-@ComponentScan(basePackages = "ch.unisg.ics.edpo")
-public class BankApplication {
-
-
-	public static void main(String[] args) {
-		SpringApplication.run(BankApplication.class, args);
-	}
-
-
-
-}
