--- bet-it/bank/src/main/java/ch/unisg/domain/utils/BankException.java
@@ -1,7 +1,0 @@
-package ch.unisg.domain.utils;
-
-public class BankException extends Exception {
-    public BankException(String message) {
-        super(message);
-    }
-}
