+++ bet-it/bank/src/main/java/ch/unisg/ics/edpo/bank/domain/utils/BankException.java
@@ -1,0 +1,7 @@
+package ch.unisg.ics.edpo.bank.domain.utils;
+
+public class BankException extends Exception {
+    public BankException(String message) {
+        super(message);
+    }
+}
