+++ bet-it/bank/src/main/java/ch/unisg/ics/edpo/bank/service/ReplayService.java
@@ -1,0 +1,4 @@
+package ch.unisg.ics.edpo.bank.service;
+
+public class ReplayService {
+}
