--- bet-it/bank/src/main/java/ch/unisg/service/FreezeService.java
@@ -1,4 +1,0 @@
-package ch.unisg.service;
-
-public class FreezeService {
-}
