--- bet-it/shared/src/main/java/ch/unisg/ics/edpo/shared/kafka/KafkaConsumerFactory.java
@@ -1,57 +1,0 @@
-package ch.unisg.ics.edpo.shared.kafka;
-
-import lombok.extern.slf4j.Slf4j;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.core.ConsumerFactory;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.kafka.listener.KafkaListenerErrorHandler;
-import org.springframework.kafka.support.converter.StringJsonMessageConverter;
-import org.springframework.kafka.support.serializer.JsonDeserializer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Configuration
-@EnableKafka
-@Slf4j
-public class KafkaConsumerFactory {
-
-    @Value("${spring.kafka.consumer.bootstrap-servers}")
-    private String bootstrapServers;
-
-
-    public ConsumerFactory<String, Map<String, Object>> consumerHashMapFactory() {
-        Map<String, Object> config = new HashMap<>();
-        config.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        config.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, true);
-        config.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-        config.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class);
-        return new DefaultKafkaConsumerFactory<>(config, new StringDeserializer(), new JsonDeserializer<>(Map.class));
-
-    }
-
-    @Bean
-    public ConcurrentKafkaListenerContainerFactory<?, ?> kafkaListenerMapFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, Map<String, Object>> factory = new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(consumerHashMapFactory());
-        factory.setMessageConverter(new StringJsonMessageConverter());
-        factory.setBatchListener(true);
-        return factory;
-
-    }
-
-    @Bean
-    public KafkaListenerErrorHandler myTopicErrorHandler() {
-        return (m, e) -> {
-            log.error("Got an error {}", e.getMessage());
-            return "some info about the failure";
-        };
-
-    }
-}
